♻️ planning 도메인 apps/planning-agent 실제 이동 + 옛 경로 compat shim
변경 이유:
모노레포 목표는 src/yule_orchestrator 를 비우고 실행 단위를 apps/* 로
옮기는 것. 골격(packages/+apps/) 위에서 첫 "진짜 이주" 로 가장 자립적인
planning 도메인을 apps/planning-agent 로 실제 이동한다. 기존 24개 import
처(cli/discord/test)를 깨지 않도록 옛 경로는 shim 으로 남긴다.

주요 변경 사항:
- planning/ 13파일 → apps/planning-agent/src/yule_planning/ git mv(히스토리 보존).
- 이동된 파일의 parent-relative import(from ..core/.. integrations/..storage)
  18개를 yule_orchestrator.* 절대 경로로 변환. intra-package(from .X)는 유지.
- 옛 경로 src/yule_orchestrator/planning/: __init__ 은 public 20 names re-export,
  12 submodule 은 sys.modules alias(identity 보존) shim.
- 루트 pyproject where 에 apps/planning-agent/src 추가.

비고:
과도기 부채 — yule_planning 은 아직 yule_orchestrator.{core,integrations,
storage} 공유 인프라를 import(apps→monolith). 해당 인프라를 packages/* 로
추출하면 edge 소멸. app pyproject/테스트는 후속 커밋.
